+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Mega Zap ⚡️ - Change log
 
+## [0.2.0] - 2025-06-02
+
 ## [Unreleased]
+
+### Added
+
+* Sorting for contacts in all records
+* Setting sheet number via supplemental info workaround in all records
+* Setting local paper map product hierarchy level code in overall record
+
+### Changed
+
+* No longer adds Side A/B extents to overall record (as the catalogue will resolve this via aggregation) 
+* No longer adds Side A/B thumbnails to overall record (as the catalogue will resolve this via aggregation) 
+* Including purchase option distribution option in all records and not just overall record
+* Updating documentation
 
 ## [0.1.0] - 2025-05-23